requirements.txt
Flask
Pillow
torch